refactor: 데이터 계층 의존 변경
